Util folder refactoring
