+++ src/components/App.tsx
@@ -6,6 +6,7 @@ import { matchPath, Router, Switch } from 'react-router';
 import { Route } from 'react-router-dom';
 import constants from '~/constants/constants';
 import EndpointPath from '~/enums/endpointPath';
+import LocalStorageHelper from '~/helpers/LocalStorageHelper';
 import navigator from '~/routing/navigator';
 import SubSites from '~/routing/subSites';
 import AuthService, { IAuthorizationResponse } from '~/services/authService';
@@ -16,7 +17,6 @@ import { LoginStore } from '~/stores/loginStore';
 import { MapStore } from '~/stores/mapStore';
 import HttpUtils from '~/utils/HttpUtils';
 import '~/utils/KeyEventHandler';
-import LocalStorageHelper from '~/utils/LocalStorageHelper';
 import ErrorBar from './ErrorBar';
 import NavigationBar from './NavigationBar';
 import * as s from './app.scss';
